chore(ci): Make message for .github/workflows/transform more explicit
We currently have two commits (basically back to back) that say `Update
OpenAPI specs`. This makes it hard to distinguish which commit is
generated by the upstream repo (hint: it's the one that has a commit
description with various SHAs) and which is generated by
`gleanwork/open-api`'s internal workflows.