refactor: GH Pages Publishing pipeline
This PR adds a new top level directory for the fully "final" specs. The
goal here is to make a location that is stable that we can reliably use
from developers.glean.com. 

This is a fairly minor change, but paves the way for a larger set of
changes to modify the code samples. Landing this first means that we can
update gleanwork/glean-developer-site to use the new `docs/specs/final`
location for it's ingestion.